+++ src/main/java/com/anthonyhilyard/merchantmarkers/MerchantMarkers.java
@@ -4,9 +4,15 @@ import com.anthonyhilyard.merchantmarkers.compat.OptifineHandler;
 import com.anthonyhilyard.merchantmarkers.render.Markers;
 
 import org.apache.commons.lang3.exception.ExceptionUtils;
+import org.lwjgl.glfw.GLFW;
 
+import com.mojang.blaze3d.platform.InputConstants;
+
+import net.minecraft.client.KeyMapping;
+import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.event.RenderNameplateEvent;
+import net.minecraftforge.client.event.RenderNameTagEvent;
+import net.minecraftforge.client.settings.KeyConflictContext;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.server.ServerStartedEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -16,6 +22,8 @@ import net.minecraftforge.fml.loading.FMLEnvironment;
 
 public class MerchantMarkers
 {
+	private static final KeyMapping showMarkers = new KeyMapping("merchantmarkers.key.showMarkers", KeyConflictContext.IN_GAME, InputConstants.Type.KEYSYM.getOrCreate(GLFW.GLFW_KEY_LEFT_ALT), "merchantmarkers.key.categories.merchantMarkers");
+
 	public void onClientSetup(FMLClientSetupEvent event)
 	{
 		try
@@ -42,9 +50,17 @@ public class MerchantMarkers
 		}
 	}
 
-	@SubscribeEvent
-	public static void onRenderNameplate(RenderNameplateEvent event)
+	public void onRegisterKeyMappings(RegisterKeyMappingsEvent event)
 	{
-		Markers.renderMarker(event.getEntityRenderer(), event.getEntity(), event.getContent(), event.getPoseStack(), event.getMultiBufferSource(), event.getPackedLight());
+		event.register(showMarkers);
+	}
+
+	@SubscribeEvent
+	public static void onRenderNameplate(RenderNameTagEvent event)
+	{
+		if (showMarkers.isDown() || MerchantMarkersConfig.getInstance().alwaysShow.get())
+		{
+			Markers.renderMarker(event.getEntityRenderer(), event.getEntity(), event.getContent(), event.getPoseStack(), event.getMultiBufferSource(), event.getPackedLight());
+		}
 	}
 }